# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api documentation",
        "description": "Api documentation",
        "contact": {
            "email": " [email protected]"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/register": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "53542a6b874642013ed45fda4c2fabf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "c_password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string",
                                        "format": "name",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Email of the user",
                                        "type": "string",
                                        "format": "email",
                                        "example": " [email protected]"
                                    },
                                    "password": {
                                        "description": "Password of the user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "c_password": {
                                        "description": "Confirm password of the user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User register successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "User token",
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "name": {
                                            "description": "Name of the user",
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "description": "Login a user",
                "operationId": "2ae5e4fc954ce9548d7be7494c98aad0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email of the user",
                                        "type": "string",
                                        "format": "email",
                                        "example": " [email protected]"
                                    },
                                    "password": {
                                        "description": "Password of the user",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "User token",
                                            "type": "string",
                                            "example": "token"
                                        },
                                        "name": {
                                            "description": "Name of the user",
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout a user",
                "description": "Logout a user",
                "operationId": "fa8a9f737ac3eeedebed8d1411ed0e7d",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message",
                                            "type": "string",
                                            "example": "User logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "description": "Get all posts",
                "operationId": "4186b281454d4cf4c67a3254c731218a",
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PostResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "description": "Create a new post",
                "operationId": "c525736cca34af3402c290c5a680dbde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the post",
                                        "type": "string",
                                        "format": "title",
                                        "example": "Post Title"
                                    },
                                    "content": {
                                        "description": "Content of the post",
                                        "type": "string",
                                        "format": "content",
                                        "example": "Post Content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a post",
                "description": "Get a post",
                "operationId": "e0f9a2214074a618841855b152721330",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post",
                "description": "Update a post",
                "operationId": "0dec9d9e95078da66ec7717d5b63e8f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the post",
                                        "type": "string",
                                        "format": "title",
                                        "example": "Post Title"
                                    },
                                    "content": {
                                        "description": "Content of the post",
                                        "type": "string",
                                        "format": "content",
                                        "example": "Post Content"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PostResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "description": "Delete a post",
                "operationId": "11cb021c9991dd024d16206da5df0ecd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a user",
                "description": "Get a user",
                "operationId": "40a5104e23b52137e78c3d3ac48b88a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PostResource": {
                "title": "PostResource",
                "properties": {
                    "id": {
                        "description": "ID of the post",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the post",
                        "type": "string"
                    },
                    "content": {
                        "description": "Content of the post",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID of the user who created the post",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Date and time the post was created",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "Date and time the post was updated",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post",
                "description": "Post model",
                "properties": {
                    "id": {
                        "description": "Post ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Post title",
                        "type": "string"
                    },
                    "content": {
                        "description": "Post content",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Post created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Post updated at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "User email verified at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "User created at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "User updated at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Everything about authentication"
        },
        {
            "name": "Users",
            "description": "Everything about users"
        },
        {
            "name": "Posts",
            "description": "Everything about posts"
        }
    ]
}